// packages/remote/src/enums/protocol/rpc-proof-operation-kind.enum.ts
/**
 * @overview Built-in RPC proof operation kind enum.
 * @author AEPKILL
 * @created 2026-08-20 00:00:00
 */

export enum RpcProofOperationKindEnum {
	freshAccept = "fresh-accept",
	resumeRequest = "resume-request",
	resumeAccept = "resume-accept",
	resumeReject = "resume-reject",
	genericReject = "generic-reject",
}